+++ lib/oidc/types/options.ts
@@ -30,11 +30,12 @@ export interface TokenParams extends CustomUrls {
   pkce?: boolean;
   clientId?: string;
   redirectUri?: string;
-  responseType?: OAuthResponseType | OAuthResponseType[];
+  responseType?: OAuthResponseType | OAuthResponseType[] | 'none';
   responseMode?: OAuthResponseMode;
   state?: string;
   nonce?: string;
   scopes?: string[];
+  enrollAmrValues?: string | string[];
   display?: string;
   ignoreSignature?: boolean;
   codeVerifier?: string;
@@ -64,6 +65,11 @@ export interface TokenManagerOptions {
   storageKey?: string;
   expireEarlySeconds?: number;
   syncStorage?: boolean;
+}
+
+export interface EnrollAuthenticatorOptions extends TokenParams {
+  enrollAmrValues: string | string[];
+  acrValues: string;
 }
 
 export interface SigninWithRedirectOptions extends TokenParams {