+++ test/spec/OktaAuth/browser.ts
@@ -482,6 +482,13 @@ describe('OktaAuth (browser)', function() {
       await auth.storeTokensFromRedirect();
       expect(auth.tokenManager.setTokens).toHaveBeenCalledWith({ accessToken, idToken });
     });
+    it('does not store tokens if responseType is "none"', async () => {
+      auth.token.parseFromUrl = jest.fn().mockResolvedValue({
+        responseType: 'none'
+      });
+      await auth.storeTokensFromRedirect();
+      expect(auth.tokenManager.setTokens).not.toHaveBeenCalled();
+    });
   });
 
   describe('setOriginalUri', () => {
@@ -725,6 +732,22 @@ describe('OktaAuth (browser)', function() {
       expect(auth.authStateManager.updateAuthState).toHaveBeenCalled();
     });
 
+  });
+
+  describe('handleRedirect', () => {
+    beforeEach(() => {
+      jest.spyOn(auth, 'handleLoginRedirect');
+    });
+
+    it('calls handleLoginRedirect', async () => {
+      await auth.handleRedirect();
+      expect(auth.handleLoginRedirect).toHaveBeenCalledWith(undefined, undefined);
+    });
+
+    it('calls handleLoginRedirect and passes originalUri', async () => {
+      await auth.handleRedirect('/overridden');
+      expect(auth.handleLoginRedirect).toHaveBeenCalledWith(undefined, '/overridden');
+    });
   });
 
 });